==> R scripts/expt_2/01_generate-trials.R <==
library(fs)
library(here)
library(readr)
library(dplyr)
library(purrr)
library(tidyr)
library(stringr)
library(jsonlite)
library(assertr)
library(magrittr)

set.seed(2021)

img_tbl <- read_csv(here::here("R Scripts", "image_table.csv"))
keys <- read_csv(here::here("R Scripts", "keys.csv"))

# design pre-randomized trial order and condition split (within-participant)

fillers <- tibble(
  domain = c("Filler-Color-red", "Filler-Shape-triangle"),
  label1 = c("gelder", "panzet"),
  label2 = ""
)

nonsense_labels <- c(
  "kapsin",
  "tantol",
  "garlet",
  "crittum",
  "tazzai",
  "pitten",
  "mipen",
  "kalmick",
  "gavai",
  "flippet",
  "belpid",
  "spadil",
  "tuffee",
  "blicket",
  "clazzo",
  "dalkeet"
)

cond_number <- c("one", "three")
cond_target <- c("label1", "label2")

group_A <- tibble(
  number = cond_number[c(1, 2, 1, 2)],
  target = cond_target[c(1, 2, 2, 1)],
  group = "A"
)
group_B <- tibble(
  number = cond_number[c(1, 2, 1, 2)],
  target = cond_target[c(2, 1, 1, 2)],
  group = "B"
)
group_C <- tibble(
  number = cond_number[c(2, 1, 2, 1)],
  target = cond_target[c(1, 2, 2, 1)],
  group = "C"
)
group_D <- tibble(
  number = cond_number[c(2, 1, 2, 1)],
  target = cond_target[c(2, 1, 1, 2)],
  group = "D"
)

cond_design <- bind_rows(group_A, group_B, group_C, group_D) %>% 
  group_by(group) %>% 
  slice(c(1:n(), 1:n())) %>% 
  mutate(id = row_number()) %>% 
  ungroup()

# cond_design %>% 
#   pivot_wider(names_from = group, values_from = c(number, target)) %>% 
#   select(1, !!!str_order(str_extract(colnames(.)[-1], "\\w$")) + 1)
# 
# | id|number_A |target_A |number_B |target_B |number_C |target_C |number_D |target_D |
# |--:|:--------|:--------|:--------|:--------|:--------|:--------|:--------|:--------|
# |  1|one      |label1   |one      |label2   |three    |label1   |three    |label2   |
# |  2|three    |label2   |three    |label1   |one      |label2   |one      |label1   |
# |  3|one      |label2   |one      |label1   |three    |label2   |three    |label1   |
# |  4|three    |label1   |three    |label2   |one      |label1   |one      |label2   |
# |  5|one      |label1   |one      |label2   |three    |label1   |three    |label2   |
# |  6|three    |label2   |three    |label1   |one      |label2   |one      |label1   |
# |  7|one      |label2   |one      |label1   |three    |label2   |three    |label1   |
# |  8|three    |label1   |three    |label2   |one      |label1   |one      |label2   |


item_design <- tibble(
  domain = setdiff(unique(img_tbl$domain), "planet"),
  label1 = nonsense_labels[c(TRUE, FALSE)],
  label2 = nonsense_labels[c(FALSE, TRUE)],
) %>%
  mutate(id = row_number())

group_designs <- cond_design %>% 
  left_join(item_design, by = "id") %>% 
  select(-id) %>% 
  group_split(group) %>%
  map_dfr(
    ~ .x %>%
      add_row(fillers[1, ], .after = 3) %>%
      add_row(fillers[2, ], .after = 7) %>%
      fill(group) %>%
      replace_na(list(target = "label1", number = "one"))
  )

## fill design with trial templates

category_dict <- img_tbl %>%
  distinct(domain, type, category) %>%
  filter(type %in% c("sub", "contrast")) %>%
  arrange(domain, desc(type))

gen_learn_set <- function(d, num_cond) {
  if (str_detect(d, "^Filler")) {
    if (str_detect(d, "Shape")) {
      "Fillers/Shape/triangle.jpg"
    } else if (str_detect(d, "Color")) {
      "Fillers/Color/red-rect.jpg"
    }
  } else {
    learn_set <- category_dict %>%
      filter(domain == d) %>%
      left_join(
        filter(img_tbl, c("one" = 1, "three" = 3)[num_cond] == .data$number), # single exemplar in learn phase
        by = c("domain", "type", "category")
      )
    pull(learn_set, path)
  }
}

gen_test_set <- function(d) {
  if (str_detect(d, "^Filler")) {
    filler_dir <- paste0("image_stimuli/Fillers/", str_extract(d, "Color|Shape"), "/")
    if (str_detect(d, "Shape")) {
      c(
        "Fillers/Shape/lines-11.jpg", "Fillers/Shape/lines-13.jpg", "Fillers/Shape/triangle-5.jpg",
        "Fillers/Shape/lines-5.jpg", "Fillers/Shape/lines-3.jpg", "Fillers/Shape/lines-2.jpg",
        "Fillers/Shape/lines-9.jpg", "Fillers/Shape/lines-8.jpg", "Fillers/Shape/triangle-3.jpg",
        "Fillers/Shape/triangle-2.jpg", "Fillers/Shape/triangle-4.jpg", "Fillers/Shape/lines-7.jpg",
        "Fillers/Shape/lines-10.jpg", "Fillers/Shape/lines-6.jpg", "Fillers/Shape/lines-4.jpg",
        "Fillers/Shape/lines-1.jpg", "Fillers/Shape/triangle-1.jpg", "Fillers/Shape/lines-12.jpg"
      )
    } else if (str_detect(d, "Color")) {
      c(
        "Fillers/Color/blue-rect.jpg", "Fillers/Color/red-rect.jpg", "Fillers/Color/green-rect.jpg",
        "Fillers/Color/green-rect.jpg", "Fillers/Color/red-rect.jpg", "Fillers/Color/green-rect.jpg",
        "Fillers/Color/green-rect.jpg", "Fillers/Color/red-rect.jpg", "Fillers/Color/blue-rect.jpg",
        "Fillers/Color/blue-rect.jpg", "Fillers/Color/green-rect.jpg", "Fillers/Color/green-rect.jpg",
        "Fillers/Color/red-rect.jpg", "Fillers/Color/blue-rect.jpg", "Fillers/Color/blue-rect.jpg",
        "Fillers/Color/blue-rect.jpg", "Fillers/Color/red-rect.jpg", "Fillers/Color/red-rect.jpg"
      )
    }
  } else {
    other_refs <- img_tbl %>%
      filter(domain != d, type == "sup") %>%
      group_by(domain) %>%
      slice_sample(n = 1) %>%
      pull(path)
    domain_refs <- img_tbl %>%
      filter(domain == d, !is.na(id)) %>%
      filter(id <= 3) %>%
      pull(path)
    sample(c(other_refs, domain_refs))
  }
}

trial_template_tbl <- group_designs %>%
  rowwise() %>%
  mutate(
    learn_set = toJSON(gen_learn_set(domain, number)),
    test_set = toJSON(gen_test_set(domain))
  ) %>%
  ungroup()

write_csv(trial_template_tbl, here::here("R Scripts", "expt_2", "01_trial_templates.csv"))
